# Use one-shot HMACSHA1.HashData() for Identity TOTP values