--- a/src/Identity/test/Identity.FunctionalTests/Pages/Account/Manage/EnableAuthenticator.cs
+++ b/src/Identity/test/Identity.FunctionalTests/Pages/Account/Manage/EnableAuthenticator.cs
@@ -1,14 +1,9 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System;
-using System.Collections.Generic;
 using System.Globalization;
 using System.Net.Http;
-using System.Security.Cryptography;
-using System.Threading.Tasks;
 using AngleSharp.Dom.Html;
-using Xunit;
 
 namespace Microsoft.AspNetCore.Identity.FunctionalTests.Account.Manage
 {
@@ -50,10 +45,10 @@ namespace Microsoft.AspNetCore.Identity.FunctionalTests.Account.Manage
 
         public static string ComputeCode(string key)
         {
-            var hash = new HMACSHA1(Base32.FromBase32(key));
-            var unixTimestamp = Convert.ToInt64(Math.Round((DateTime.UtcNow - new DateTime(1970, 1, 1, 0, 0, 0)).TotalSeconds));
+            var keyBytes = Base32.FromBase32(key);
+            var unixTimestamp = DateTimeOffset.UtcNow.ToUnixTimeSeconds();
             var timestep = Convert.ToInt64(unixTimestamp / 30);
-            var topt = Rfc6238AuthenticationService.ComputeTotp(hash, (ulong)timestep, modifier: null);
+            var topt = Rfc6238AuthenticationService.ComputeTotp(keyBytes, (ulong)timestep, modifier: null);
             return topt.ToString("D6", CultureInfo.InvariantCulture);
         }
     }